+++ packages/storefront/tests/vrt/config/jest.setup.ts
@@ -3,4 +3,6 @@ import { setCustomOptions } from '@porsche-design-system/shared/testing';
 setCustomOptions({
   baseUrl: 'http://localhost:8080',
   mode: 'square-auto',
+  fixturesDir: 'tests/vrt/fixtures',
+  resultsDir: 'tests/vrt/results',
 });