+++ packages/storefront/tests/vrt/specs/markdown.vrt.ts
@@ -1,6 +1,6 @@
 import { defaultViewports, getVisualRegressionTester, vrtTest } from '@porsche-design-system/shared/testing';
 import { routerViewSelector } from '../helpers';
-import { forceFocusState } from '../../../../components-js/tests/vrt/helpers';
+import { forceFocusState } from '../../../../components-js/tests/vrt/puppeteer/helpers';
 
 it.each(defaultViewports)('should have no visual regression for viewport %s', async (viewport) => {
   expect(