+++ packages/components/CHANGELOG.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### Fixed
 
-- `SelectWrapper` dropdown keeps disabled attribute now in sync with native selection if changed programmatically
+- `Select Wrapper` with custom dropdown keeps attribute changes of native select options in sync if changed programmatically
 
 ### [2.16.1] - 2022-09-09