+++ packages/components/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### Fixed
 
+- Issue with `Popover` where drop-shadow is not shown correctly in Chrome >= 105
 - `Select Wrapper` with custom dropdown keeps attribute changes of native select options in sync if changed programmatically
 
 ### [2.16.1] - 2022-09-09