+++ packages/components/CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Issue with `Popover` where drop-shadow is not shown correctly in Chrome >= 105
 - Issue with `Carousel` and `wrap-content="true"` where the layout was out of sync with `Content Wrapper` for
   viewports >= 1760px.
+- `Select Wrapper` with custom dropdown keeps attribute changes of native select options in sync if changed programmatically
 
 ### [2.16.1] - 2022-09-09